beautiful output on test client
